% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_env_or_stop}
\alias{get_env_or_stop}
\title{Get an environment variable or fail with a clear error}
\usage{
get_env_or_stop(name)
}
\arguments{
\item{name}{Name of the environment variable.}
}
\value{
The value of the environment variable as a character scalar.
}
\description{
Get an environment variable or fail with a clear error
}
\keyword{internal}
