% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorstore_interface.R
\name{vectorstore_clear_all}
\alias{vectorstore_clear_all}
\title{Delete ALL collections from the vector store}
\usage{
vectorstore_clear_all()
}
\value{
Invisibly, TRUE
}
\description{
Completely wipes the vector store by replacing it with an empty tibble.
}
